# Init Google Tag Manager script