### Purpose

Add Tag Manager script back in so we can start tracking events

### To Test
Add `REACT_APP_GTM_ID='GTM_ID_GOES_HERE'` to your `.env` file. Feel free to ask me what the ID is and how to properly test that the triggers are working

### Performance Testing

As a test, I reloaded both `/linodes/${id}/summary` and `/linodes` 5 times with the Tag Manager script included and 5 times without. These are the results:

On Linodes Landing page:
* page load times ranged from 2572ms to 3522ms with the Google Tag Manager script
* page load times ranged from 2183ms to 3173ms without the Google Tag Manager script

On Linodes Detail page:
* page load times ranged from 2325ms to 3682ms with the Google Tag Manager script
* page load times ranged from 2303ms to 2725ms without the Google Tag Manager script

Also note that when the code to trigger a change to the `dataLayer` on the `then` and `catch` of async requests, the page loads didn't appear that much worse, but I commented out the code because until we start using Tag Manager for async actions, there isn't much need to send request data to GTM